Reorganize Tests under Handler folder
